requirements.txt
Flask
imageio[ffmpeg]
Pillow